Split yaru_progress_indicator in multiple files
