+++ lib/src/controls/yaru_progress_indicator.dart
@@ -1,16 +1,11 @@
-import 'dart:math' as math;
-
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
-const double _kMinCircularProgressIndicatorSize = 36.0;
-const double _kDefaultCircularProgressStrokeWidth = 6;
-const double _kDefaultLinearProgressMinHeight = 6;
-const int _kIndeterminateAnimationDuration = 8000;
-const Curve _kIndeterminateAnimationCurve =
+const int kIndeterminateAnimationDuration = 8000;
+const Curve kIndeterminateAnimationCurve =
     Cubic(.35, .75, .65, .25); // Kind of `Curves.slowMiddle` curve
 
-abstract class _YaruProgressIndicator extends StatefulWidget {
+abstract class YaruProgressIndicator extends StatefulWidget {
   /// Creates a Yaru progress indicator.
   ///
   /// {@template yaru.widget.YaruProgressIndicator.YaruProgressIndicator}
@@ -24,7 +19,7 @@ abstract class _YaruProgressIndicator extends StatefulWidget {
   /// bar for screen reading software. The [semanticsValue] property may be used
   /// for determinate progress indicators to indicate how much progress has been made.
   /// {@endtemplate}
-  const _YaruProgressIndicator({
+  const YaruProgressIndicator({
     super.key,
     this.value,
     this.color,
@@ -78,7 +73,7 @@ abstract class _YaruProgressIndicator extends StatefulWidget {
   /// become '10%'.
   final String? semanticsValue;
 
-  Color _getValueColor(BuildContext context) {
+  Color getValueColor(BuildContext context) {
     return valueColor?.value ??
         color ??
         ProgressIndicatorTheme.of(context).color ??
@@ -98,7 +93,7 @@ abstract class _YaruProgressIndicator extends StatefulWidget {
     );
   }
 
-  Widget _buildSemanticsWrapper({
+  Widget buildSemanticsWrapper({
     required BuildContext context,
     required Widget child,
   }) {
@@ -111,532 +106,5 @@ abstract class _YaruProgressIndicator extends StatefulWidget {
       value: expandedSemanticsValue,
       child: child,
     );
-  }
-}
-
-class YaruLinearProgressIndicator extends _YaruProgressIndicator {
-  /// Creates a Yaru linear progress indicator.
-  ///
-  /// {@macro yaru.widget.YaruProgressIndicator.YaruProgressIndicator}
-  const YaruLinearProgressIndicator({
-    super.key,
-    super.value,
-    this.minHeight = _kDefaultLinearProgressMinHeight,
-    super.color,
-    super.valueColor,
-    super.semanticsLabel,
-    super.semanticsValue,
-  }) : assert(minHeight > 0);
-
-  /// The minimum height of the line used to draw the linear indicator (default: 6).
-  final double minHeight;
-
-  @override
-  _YaruLinearProgressIndicatorState createState() =>
-      _YaruLinearProgressIndicatorState();
-}
-
-class _YaruLinearProgressIndicatorState
-    extends State<YaruLinearProgressIndicator>
-    with SingleTickerProviderStateMixin {
-  static final Animatable<double> _positionTween = Tween(begin: 0.0, end: 6.0)
-      .chain(CurveTween(curve: _kIndeterminateAnimationCurve));
-  static final Animatable<double> _speedTween = Tween(begin: -1.0, end: 1.0)
-      .chain(CurveTween(curve: _kIndeterminateAnimationCurve));
-
-  late AnimationController _controller;
-
-  @override
-  void initState() {
-    super.initState();
-    _controller = AnimationController(
-      duration: const Duration(milliseconds: _kIndeterminateAnimationDuration),
-      vsync: this,
-    );
-
-    if (widget.value == null) {
-      _controller.repeat();
-    }
-  }
-
-  @override
-  void dispose() {
-    _controller.dispose();
-    super.dispose();
-  }
-
-  @override
-  void didUpdateWidget(YaruLinearProgressIndicator oldWidget) {
-    super.didUpdateWidget(oldWidget);
-    if (widget.value == null && !_controller.isAnimating) {
-      _controller.repeat();
-    } else if (widget.value != null && _controller.isAnimating) {
-      _controller.stop();
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    if (widget.value != null) {
-      return _buildContainer(
-        context,
-        CustomPaint(
-          painter: _DeterminateYaruLinearProgressIndicatorPainter(
-            widget.value!,
-            widget._getValueColor(context),
-            Directionality.of(context),
-          ),
-        ),
-      );
-    }
-
-    return AnimatedBuilder(
-      animation: _controller,
-      builder: (context, child) {
-        return _buildContainer(
-          context,
-          ClipRRect(
-            borderRadius: BorderRadius.all(Radius.circular(widget.minHeight)),
-            child: CustomPaint(
-              painter: _IndeterminateYaruLinearProgressIndicatorPainter(
-                widget._getValueColor(context),
-                _positionTween.evaluate(_controller),
-                _speedTween.evaluate(_controller).abs(),
-                Directionality.of(context),
-              ),
-            ),
-          ),
-        );
-      },
-    );
-  }
-
-  Widget _buildContainer(BuildContext context, Widget child) {
-    return widget._buildSemanticsWrapper(
-      context: context,
-      child: Container(
-        constraints: BoxConstraints(
-          minWidth: double.infinity,
-          minHeight: widget.minHeight,
-        ),
-        child: child,
-      ),
-    );
-  }
-}
-
-class _IndeterminateYaruLinearProgressIndicatorPainter extends CustomPainter {
-  const _IndeterminateYaruLinearProgressIndicatorPainter(
-    this.color,
-    this.position,
-    this.speed,
-    this.textDirection,
-  ) : super();
-
-  final Color color;
-  final double position;
-  final double speed;
-  final TextDirection textDirection;
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    final realPosition = position - position.truncate();
-    final strokeWidth = size.width / 6;
-    final y = size.height / 2;
-
-    final fillPaint = Paint()
-      ..color = color
-      ..style = PaintingStyle.fill;
-
-    if (textDirection == TextDirection.rtl) {
-      canvas.scale(-1, 1);
-      canvas.translate(-size.width, 0);
-    }
-
-    for (var i = 0; i < 3; i++) {
-      final x1 = size.width * realPosition +
-          (strokeWidth * i / 2 + (strokeWidth * i * speed));
-      final x2 = size.width * realPosition +
-          strokeWidth +
-          (strokeWidth * i / 2 + (strokeWidth * i * speed));
-
-      canvas.drawLine(
-        Offset(x1, y),
-        Offset(x2, y),
-        _getGradientPaint(color, x1, strokeWidth, size.height, speed),
-      );
-      canvas.drawCircle(Offset(x2, y), size.height / 2, fillPaint);
-
-      // If a line overflow on the right, redraw it on the left
-      if (x2 > size.width) {
-        canvas.drawLine(
-          Offset(x1 - size.width, y),
-          Offset(x2 - size.width, y),
-          _getGradientPaint(
-            color,
-            x1 - size.width,
-            strokeWidth,
-            size.height,
-            speed,
-          ),
-        );
-        canvas.drawCircle(
-          Offset(x2 - size.width, y),
-          size.height / 2,
-          fillPaint,
-        );
-      }
-    }
-  }
-
-  Paint _getGradientPaint(
-    Color color,
-    double x,
-    double width,
-    double height,
-    double speed,
-  ) {
-    final gradient = LinearGradient(
-      colors: [
-        color.withAlpha(0),
-        color.withAlpha(20 + (230 * speed).toInt()),
-        color.withAlpha(150 + (100 * speed).toInt())
-      ],
-      stops: const [0.15, 0.75, 1],
-    );
-
-    return Paint()
-      ..shader = gradient.createShader(Rect.fromLTWH(x, 0.0, width, height))
-      ..strokeWidth = height
-      ..strokeCap = StrokeCap.round
-      ..style = PaintingStyle.stroke;
-  }
-
-  @override
-  bool shouldRepaint(
-    _IndeterminateYaruLinearProgressIndicatorPainter oldDelegate,
-  ) {
-    return oldDelegate.color != color ||
-        oldDelegate.position != position ||
-        oldDelegate.speed != speed ||
-        oldDelegate.textDirection != textDirection;
-  }
-}
-
-class _DeterminateYaruLinearProgressIndicatorPainter extends CustomPainter {
-  const _DeterminateYaruLinearProgressIndicatorPainter(
-    this.value,
-    this.color,
-    this.textDirection,
-  ) : super();
-
-  final double value;
-  final Color color;
-  final TextDirection textDirection;
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    final revisedValue = value >= 0 && value <= 1
-        ? value
-        : value < 0
-            ? 0
-            : 1;
-    final candidateBackgroundHeight = (size.height / 3 * 2).truncate();
-    final backgroundHeight =
-        (candidateBackgroundHeight + (candidateBackgroundHeight.isEven ? 0 : 1))
-            .toDouble();
-
-    final backgroundPaint = Paint()
-      ..color = color.withOpacity(.25)
-      ..strokeWidth = backgroundHeight
-      ..strokeCap = StrokeCap.round
-      ..style = PaintingStyle.stroke;
-    final strokePaint = Paint()
-      ..color = color
-      ..strokeWidth = size.height
-      ..strokeCap = StrokeCap.round
-      ..style = PaintingStyle.stroke;
-
-    if (textDirection == TextDirection.rtl) {
-      canvas.scale(-1, 1);
-      canvas.translate(-size.width, 0);
-    }
-
-    final y = size.height / 2;
-
-    canvas.drawLine(
-      Offset(backgroundHeight / 2, y),
-      Offset(size.width - backgroundHeight / 2, y),
-      backgroundPaint,
-    );
-
-    if (size.width * revisedValue > size.height) {
-      canvas.drawLine(
-        Offset(y, y),
-        Offset((size.width - y) * revisedValue, y),
-        strokePaint,
-      );
-    } else if (revisedValue > 0) {
-      final fillPaint = Paint()
-        ..color = color
-        ..style = PaintingStyle.fill;
-
-      canvas.drawCircle(Offset(y, y), y, fillPaint);
-    }
-  }
-
-  @override
-  bool shouldRepaint(
-    _DeterminateYaruLinearProgressIndicatorPainter oldDelegate,
-  ) {
-    return oldDelegate.value != value ||
-        oldDelegate.color != color ||
-        oldDelegate.textDirection != textDirection;
-  }
-}
-
-class YaruCircularProgressIndicator extends _YaruProgressIndicator {
-  /// Creates a Yaru circular progress indicator.
-  ///
-  /// {@macro yaru.widget.YaruProgressIndicator.YaruProgressIndicator}
-  const YaruCircularProgressIndicator({
-    super.key,
-    super.value,
-    this.strokeWidth = _kDefaultCircularProgressStrokeWidth,
-    super.color,
-    super.valueColor,
-    super.semanticsLabel,
-    super.semanticsValue,
-  });
-
-  /// The width of the line used to draw the circle.
-  final double strokeWidth;
-
-  @override
-  _YaruCircularProgressIndicatorState createState() =>
-      _YaruCircularProgressIndicatorState();
-}
-
-class _YaruCircularProgressIndicatorState
-    extends State<YaruCircularProgressIndicator>
-    with SingleTickerProviderStateMixin {
-  static final Animatable<double> _rotationTween =
-      Tween(begin: 0.0, end: math.pi * 12)
-          .chain(CurveTween(curve: _kIndeterminateAnimationCurve));
-  static final Animatable<double> _speedTween = Tween(begin: -1.0, end: 1.0)
-      .chain(CurveTween(curve: _kIndeterminateAnimationCurve));
-
-  late AnimationController _controller;
-
-  @override
-  void initState() {
-    super.initState();
-    _controller = AnimationController(
-      duration: const Duration(milliseconds: _kIndeterminateAnimationDuration),
-      vsync: this,
-    );
-
-    if (widget.value == null) {
-      _controller.repeat();
-    }
-  }
-
-  @override
-  void dispose() {
-    _controller.dispose();
-    super.dispose();
-  }
-
-  @override
-  void didUpdateWidget(YaruCircularProgressIndicator oldWidget) {
-    super.didUpdateWidget(oldWidget);
-    if (widget.value == null && !_controller.isAnimating) {
-      _controller.repeat();
-    } else if (widget.value != null && _controller.isAnimating) {
-      _controller.stop();
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    if (widget.value != null) {
-      return _buildContainer(
-        context,
-        CustomPaint(
-          painter: _DeterminateYaruCircularProgressIndicatorPainter(
-            widget.value!,
-            widget._getValueColor(context),
-            widget.strokeWidth,
-            Directionality.of(context),
-          ),
-        ),
-      );
-    }
-    return AnimatedBuilder(
-      animation: _controller,
-      builder: (context, child) {
-        return _buildContainer(
-          context,
-          CustomPaint(
-            painter: _IndeterminateYaruCircularProgressIndicatorPainter(
-              widget._getValueColor(context),
-              widget.strokeWidth,
-              _rotationTween.evaluate(_controller),
-              _speedTween.evaluate(_controller).abs(),
-              Directionality.of(context),
-            ),
-          ),
-        );
-      },
-    );
-  }
-
-  Widget _buildContainer(BuildContext context, Widget child) {
-    return widget._buildSemanticsWrapper(
-      context: context,
-      child: Container(
-        constraints: const BoxConstraints(
-          minWidth: _kMinCircularProgressIndicatorSize,
-          minHeight: _kMinCircularProgressIndicatorSize,
-        ),
-        child: child,
-      ),
-    );
-  }
-}
-
-class _IndeterminateYaruCircularProgressIndicatorPainter extends CustomPainter {
-  const _IndeterminateYaruCircularProgressIndicatorPainter(
-    this.color,
-    this.strokeWidth,
-    this.rotationAngle,
-    this.speed,
-    this.textDirection,
-  ) : super();
-
-  final Color color;
-  final double strokeWidth;
-  final double rotationAngle;
-  final double speed;
-  final TextDirection textDirection;
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    const circleThird = math.pi * 2 / 3;
-
-    final gradient = SweepGradient(
-      startAngle: 0.0 + rotationAngle,
-      endAngle: circleThird + rotationAngle,
-      tileMode: TileMode.repeated,
-      colors: [
-        color.withAlpha(0),
-        color.withAlpha(20 + (230 * speed).toInt()),
-        color.withAlpha(250)
-      ],
-      stops: const [0.15, 0.75, 1],
-    );
-
-    final center = Offset(size.width / 2, size.height / 2);
-    final radius =
-        math.min(size.width / 2, size.height / 2) - (strokeWidth / 2);
-    const sweepAngle = circleThird;
-    final rect = Rect.fromCircle(center: center, radius: radius);
-
-    final gradientPaint = Paint()
-      ..shader = gradient.createShader(rect)
-      ..style = PaintingStyle.stroke
-      ..strokeWidth = strokeWidth;
-    final fillPaint = Paint()
-      ..color = color
-      ..style = PaintingStyle.fill;
-
-    if (textDirection == TextDirection.rtl) {
-      canvas.scale(-1, 1);
-      canvas.translate(-size.width, 0);
-    }
-
-    for (var i = 0; i < 3; i++) {
-      final startAngle = rotationAngle + circleThird * i;
-      canvas.drawArc(rect, startAngle, sweepAngle, false, gradientPaint);
-    }
-
-    // Draw circles after arcs, so they look on top
-    for (var i = 0; i < 3; i++) {
-      final startAngle = rotationAngle + circleThird * i;
-      canvas.drawCircle(
-        Offset(
-          math.cos(startAngle) * radius + center.dx,
-          math.sin(startAngle) * radius + center.dy,
-        ),
-        strokeWidth / 2,
-        fillPaint,
-      );
-    }
-  }
-
-  @override
-  bool shouldRepaint(
-    _IndeterminateYaruCircularProgressIndicatorPainter oldDelegate,
-  ) {
-    return oldDelegate.color != color ||
-        oldDelegate.strokeWidth != strokeWidth ||
-        oldDelegate.rotationAngle != rotationAngle ||
-        oldDelegate.speed != speed ||
-        oldDelegate.textDirection != textDirection;
-  }
-}
-
-class _DeterminateYaruCircularProgressIndicatorPainter extends CustomPainter {
-  const _DeterminateYaruCircularProgressIndicatorPainter(
-    this.value,
-    this.color,
-    this.width,
-    this.textDirection,
-  ) : super();
-
-  final double value;
-  final Color color;
-  final double width;
-  final TextDirection textDirection;
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    final revisedValue = value >= 0 && value <= 1
-        ? value
-        : value < 0
-            ? 0
-            : 1;
-    final center = Offset(size.width / 2, size.height / 2);
-    final radius = math.min(size.width / 2, size.height / 2) - (width / 2);
-    const startAngle = -math.pi / 2;
-    final sweepAngle = math.pi * 2 * revisedValue;
-    final rect = Rect.fromCircle(center: center, radius: radius);
-
-    if (textDirection == TextDirection.rtl) {
-      canvas.scale(-1, 1);
-      canvas.translate(-size.width, 0);
-    }
-
-    final backgroundPaint = Paint()
-      ..color = color.withOpacity(.25)
-      ..strokeWidth = width > 2 ? width - 2 : width
-      ..style = PaintingStyle.stroke;
-    final strokePaint = Paint()
-      ..color = color
-      ..strokeWidth = width
-      ..style = PaintingStyle.stroke;
-
-    canvas.drawArc(rect, 0, math.pi * 2, false, backgroundPaint);
-    canvas.drawArc(rect, startAngle, sweepAngle, false, strokePaint);
-  }
-
-  @override
-  bool shouldRepaint(
-    _DeterminateYaruCircularProgressIndicatorPainter oldDelegate,
-  ) {
-    return oldDelegate.value != value ||
-        oldDelegate.color != color ||
-        oldDelegate.width != width ||
-        oldDelegate.textDirection != textDirection;
   }
 }
