+++ lib/yaru_widgets.dart
@@ -6,12 +6,13 @@ export 'src/constants.dart';
 export 'src/controls/yaru_back_button.dart';
 export 'src/controls/yaru_check_button.dart';
 export 'src/controls/yaru_checkbox.dart';
+export 'src/controls/yaru_circular_progress_indicator.dart';
 export 'src/controls/yaru_close_button.dart';
 export 'src/controls/yaru_color_disk.dart';
 export 'src/controls/yaru_icon_button.dart';
+export 'src/controls/yaru_linear_progress_indicator.dart';
 export 'src/controls/yaru_option_button.dart';
 export 'src/controls/yaru_popup_menu_button.dart';
-export 'src/controls/yaru_progress_indicator.dart';
 export 'src/controls/yaru_radio.dart';
 export 'src/controls/yaru_radio_button.dart';
 export 'src/controls/yaru_switch.dart';
